+++ Commands/about.js
@@ -1,9 +1,8 @@
-const Command = require('./../Classes/Command.js');
-const { getString } = require('./../Localizations/localization.js');
+const Command = require('../Classes/Command.js');
+const { getString } = require('../Localizations/localization.js');
 const { MessageEmbed } = require('discord.js');
-const { versionBuilder } = require('../helpers.js');
 
-var command = new Command("credits", false, false, true);
+var command = new Command("about", false, false, true);
 
 // Overwrite detailed help description with executing the command
 command.help = (avatarURL, guildID, locale, guildName, module) => {
@@ -14,8 +13,6 @@ command.execute = (receivedMessage, state, locale) => {
 	// Displays the credits
 	receivedMessage.author.send(creditsBuilder(receivedMessage.client.user.avatarURL(), locale))
 		.catch(console.error);
-	receivedMessage.author.send(versionBuilder(receivedMessage.client.user.avatarURL()))
-		.catch(console.error);
 }
 
 module.exports = command;
@@ -23,7 +20,7 @@ module.exports = command;
 function creditsBuilder(footerURL, locale) {
 	return new MessageEmbed().setColor(`6b81eb`)
 		.setAuthor("Imaginary Horizons Productions", `https://cdn.discordapp.com/icons/353575133157392385/c78041f52e8d6af98fb16b8eb55b849a.png `, `https://discord.gg/bcE3Syu `)
-		.setTitle(getString(locale, command.module, "creditsTitle").addVariables({ "versionNumber": "2.1.0" }))
+		.setTitle(getString(locale, command.module, "creditsTitle"))
 		.setThumbnail('https://cdn.discordapp.com/attachments/545684759276421120/734097014974971924/theater-curtains.png')
 		.setURL(`https://github.com/Imaginary-Horizons-Productions/DirectoryBot `)
 		.addField(getString(locale, command.module, "designAndEngineering"), `Nathaniel Tseng ( <@106122478715150336> | [Twitter](https://twitter.com/Archainis) )`)
@@ -31,7 +28,7 @@ function creditsBuilder(footerURL, locale) {
 		.addField(getString(locale, command.module, "art"), `Angela Lee ( [Website](https://www.angelasylee.com/) )`)
 		.addField(getString(locale, command.module, "localization"), `Mnemmy (fr), towoko (de), Lille (de)`)
 		.addField(`Embed Thumbnails`, `[game-icons.net](https://game-icons.net/)`)
-		.addField(`\u200B`, getString(locale, "credits", "patronsDescription"))
+		.addField(`\u200B`, getString(locale, command.module, "patronsDescription"))
 		.addField(getString(locale, command.module, "cartographerTier"), `Ralph Beish`, false)
 		.addField(getString(locale, command.module, "explorerTier"), `Eric Hu`, false)
 		.setFooter(getString(locale, "DirectoryBot", "footerText"), footerURL)