+++ Templates/command_template.js
@@ -1,5 +1,5 @@
-const Command = require('./../Classes/Command.js');
-const { getString } = require('./../Localizations/localization.js');
+const Command = require('../Classes/Command.js');
+const { getString } = require('../Localizations/localization.js');
 
 var command = new Command("", false, false, false);